requirements.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
requests